Fixed te structure of the folders
