+++ backend/src/main/java/com/codecool/restmates/dto/requests/NewAccommodationDTO.java
@@ -1,4 +1,4 @@
-package com.codecool.restmates.service.accommodation.DTO;
+package com.codecool.restmates.dto.requests;
 
 public record NewAccommodationDTO(long accommodation_id, String name, double price_per_night) {
 }
