+++ backend/src/main/java/com/codecool/restmates/dto/requests/NewMemberDTO.java
@@ -1,0 +1,4 @@
+package com.codecool.restmates.dto.requests;
+
+public record NewMemberDTO(String member_id, String first_name, String last_name, String email) {
+}
