+++ backend/src/main/java/com/codecool/restmates/dto/responses/AccommodationResponseDTO.java
@@ -1,4 +1,4 @@
-package com.codecool.restmates.service.accommodation.DTO;
+package com.codecool.restmates.dto.responses;
 
 public record AccommodationResponseDTO(long accommodation_id, String name, double price_per_night) {
 }
