+++ backend/src/main/java/com/codecool/restmates/dto/responses/MemberResponseDTO.java
@@ -1,0 +1,4 @@
+package com.codecool.restmates.dto.responses;
+
+public record MemberResponseDTO(String memberId, String firstName, String lastName, String email) {
+}
