--- backend/src/main/java/com/codecool/restmates/service/member/DAO/MemberDAO.java
@@ -1,12 +1,0 @@
-package com.codecool.restmates.service.member.DAO;
-
-import com.codecool.restmates.model.Member;
-import com.codecool.restmates.service.member.DTO.MemberResponseDTO;
-import com.codecool.restmates.service.member.DTO.NewMemberDTO;
-
-public interface MemberDAO {
-    MemberResponseDTO getMemberById(long id);
-    MemberResponseDTO createMember(NewMemberDTO member);
-    void deleteMemberById(long id);
-    Member updateMember(Member member);
-}
