--- backend/src/main/java/com/codecool/restmates/service/member/DTO/MemberResponseDTO.java
@@ -1,4 +1,0 @@
-package com.codecool.restmates.service.member.DTO;
-
-public record MemberResponseDTO(long member_id, String first_name, String last_name, String email) {
-}
