--- backend/src/main/java/com/codecool/restmates/service/member/DTO/NewMemberDTO.java
@@ -1,4 +1,0 @@
-package com.codecool.restmates.service.member.DTO;
-
-public record NewMemberDTO(long member_id, String first_name, String last_name, String email) {
-}
